TelegramRevamp/pyproject.toml
[project]
name = "telegram-bot-platform"
version = "1.0.0"
description = "Telegram Bot Platform for Orders Management"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.8"
readme = "PROJECT_GUIDE.md"
license = {text = "MIT"}

dependencies = [
    "aiogram>=3.22.0",
    "asyncpg>=0.31.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gdown>=5.2.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "psycopg2-binary>=2.9.9",
    "aiohttp>=3.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
